# Cythonise the keysort functions in idx-opt.c