--- a/tables/tests/test_aux.py
+++ b/tables/tests/test_aux.py
@@ -1,0 +1,34 @@
+# -*- coding: utf-8 -*-
+
+import unittest
+import numpy
+
+from tables import indexesextension
+
+
+class TestAuxiliaryFunctions(unittest.TestCase):
+    def test_keysort(self):
+        N = 1000
+        rnd = numpy.random.randint(N, size=N)
+        for dtype1 in ('S6', 'b1', 'i1', 'i8', 'u4', 'u8', 'f4', 'f8'):
+            for dtype2 in ('u4', 'i8'):
+                a = numpy.array(rnd, dtype1)
+                b = numpy.array(rnd, dtype2)
+
+                c = a.copy()
+                d = c.argsort()
+                e = c[d]
+                f = b[d]
+
+                indexesextension.keysort(a, b)
+                self.assertTrue((a == e).all())
+                self.assertTrue((b == f).all())
+
+
+def suite():
+    theSuite = unittest.TestSuite()
+    theSuite.addTest(unittest.makeSuite(TestAuxiliaryFunctions))
+    return theSuite
+
+if __name__ == '__main__':
+    unittest.main(defaultTest='suite')