… src/idx-opt.c