This is #301 with minor modifications rebased on top of the branch `release-3.2.0`.