@FrancescAlted 2015-04-18
Ok. I did some quick benchmarks, and performance seems pretty much unchanged:

Before the patch:

```
faltet@francesc-Latitude-E6430:~/software/PyTables-upstream$ PYTHONPATH=. python bench/search-bench.py -n 1000 -N 100 -w /tmp/bench.h5
Time for filling: 0.098 Krows/s: 10232.381 , File size: 15.385 MB
Time for indexing: 1.262 iKrows/s: 792.127 , Index size: 28.444 MB
Rows written: 1000000  Row size: 16
Time writing rows: 0.098 s (real) 0.098 s (cpu)  100%
Total file size: 43.829 MB , Write KB/s (pure data): 159438
Rows indexed: 1000000  (IMRows): 1.0
Time indexing rows: 1.262 s (real) 0.793 s (cpu)  63%
```

After the patch:

```
faltet@francesc-Latitude-E6430:~/software/PyTables-upstream$ PYTHONPATH=. python bench/search-bench.py -n 1000 -N 100 -w /tmp/bench.h5
Time for filling: 0.137 Krows/s: 7283.371 , File size: 15.385 MB
Time for indexing: 1.273 iKrows/s: 785.379 , Index size: 28.444 MB
Rows written: 1000000  Row size: 16
Time writing rows: 0.137 s (real) 0.137 s (cpu)  100%
Total file size: 43.829 MB , Write KB/s (pure data): 114051
Rows indexed: 1000000  (IMRows): 1.0
Time indexing rows: 1.273 s (real) 0.95 s (cpu)  75%
```

So, merging.  Thanks Andrea!