@andreabedini 2015-04-18
Thanks for merging and for the benchmark @FrancescAlted